requirements.txt
asgiref==3.2.10
beautiful==0.0.2
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.1.2
django-allauth==0.44.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-ipware==4.0.0
django-rest-auth==0.9.5
django-rest-authtoken==2.1.3
django-rest-framework==0.1.0
django-rest-knox==4.1.0
django-webpack-loader==1.0.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
idna==2.10
joblib==1.1.0
mysqlclient==2.0.3
numpy==1.21.4
oauthlib==3.1.0
pandas==1.3.4
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
scikit-learn==1.0.1
scipy==1.7.2
six==1.16.0
sklearn==0.0
soupsieve==2.3.1
sqlparse==0.4.1
threadpoolctl==3.0.0
tqdm==4.62.3
urllib3==1.26.5
